[IN-PROCESS] load 3DObject
