+++ src/D3vS1m.WebUI/ClientApp/components/templates/objFile.vue
@@ -18,7 +18,7 @@
 </template>
 
 <script>
-import utils from '../../services/ultis'
+import utils from '../../services/loadObjFile'
 export default {
     props: ['objFile'],
     methods: {
