+++ src/D3vS1m.WebUI/Views/Home/Index.cshtml
@@ -11,8 +11,4 @@
     <script src="~/dist/main.js" asp-append-version="true"></script>
     <!-- Compiled and minified JavaScript -->
     <script src="https://cdnjs.cloudflare.com/ajax/libs/materialize/1.0.0/js/materialize.min.js"></script>
-     <script src="https://threejsfundamentals.org/threejs/resources/threejs/r105/three.min.js"></script>
-<script src="https://threejsfundamentals.org/threejs/resources/threejs/r105/js/controls/OrbitControls.js"></script>
-<script src="https://threejsfundamentals.org/threejs/resources/threejs/r105/js/loaders/LoaderSupport.js"></script>
-<script src="https://threejsfundamentals.org/threejs/resources/threejs/r105/js/loaders/OBJLoader2.js"></script>
 }
